polish(db-jpa): P3 코드리뷰 6 MED + 2 LOW 반영
- NativeRowMapper.toLong/toInt: 비-Number 입력 시 IllegalStateException(toLdt와 동일 실패 모드)
- CouncilProjectRow.fromRow: r[12] NULL placeholder 불변식 가드(SELECT 순서변경 조기 검출)
- #6 DTO 4종(EvaluationItemAvgRow/PendingApprovalRow/LabeledCountRow/RecentReviewingRow): EXPECTED_COLUMNS 길이 가드 추가
- EvaluationItemAvgRow: r[1]이 BigDecimal이면 그대로 사용해 IEEE-754 변환 아티팩트 회피
- Application/ServiceRequestDocService: (int) 캐스트 → Math.toIntExact (silent narrowing 방지)
- AdminService.getLoginStats: label null/blank 필터 후 LocalDate.parse
- ProjectListProjectionIt: rprStsTc 동등성 단언 추가(abusTc/rprStsTc 위치 스왑 감지)
- 빈 목록 동등성 IT 2건: 매핑 검증 의존성 주석 명시

+++ src/main/java/com/kdb/it/common/approval/service/ApplicationService.java
@@ -550,7 +550,7 @@ public class ApplicationService {
             applicationRepository.findMonthlyTrendRowsByBbrC(bbrC).stream()
                 .map(row -> ApplicationDto.MonthlyCount.builder()
                     .month(row.label())
-                    .count((int) row.count())
+                    .count(Math.toIntExact(row.count()))
                     .build())
                 .toList();
