+++ src/main/java/com/kdb/it/domain/budget/document/service/ServiceRequestDocService.java
@@ -326,7 +326,7 @@ public class ServiceRequestDocService {
             serviceRequestDocRepository.findMonthlyTrendRowsByBbrC(bbrC).stream()
                 .map(row -> ServiceRequestDocDto.MonthlyCount.builder()
                     .month(row.label())
-                    .count((int) row.count())
+                    .count(Math.toIntExact(row.count()))
                     .build())
                 .toList();
